Add Neil to contribs
Neil worked on ant with me several months ago; taught me a LOT; and did a lot of good work. When I was re-writing ant, so much changed that I ended up having to get rid of a lot of his contributions, but he really helped me grow and deserves notice for it.